Add AppVeyor CI

+++ src/game/server/gamecontext.h
@@ -180,7 +180,7 @@ public:
 };
 
 inline int64 CmaskAll() { return -1; }
-inline int64 CmaskOne(int ClientID) { return 1<<ClientID; }
+inline int64 CmaskOne(int ClientID) { return (int64)1<<ClientID; }
 inline int64 CmaskAllExceptOne(int ClientID) { return CmaskAll()^CmaskOne(ClientID); }
 inline bool CmaskIsSet(int64 Mask, int ClientID) { return (Mask&CmaskOne(ClientID)) != 0; }
 #endif
